incorrect merge resolutions?
...or was 2788.multi_intro.0 not correct?

This patch represents differences between the head of a branch
started from leif/introless-multiintro to get correct attribution
for 2788.multi_intro.0 However, at the end there were some differences
and so this diff takes out those differences so that the tip of
this branch and 2788.multi_intro.0 are the same.

If any of the changes in this diff look wrong, they probably are.