Fix writing mutable files after leasedb loss
Previously, if the leasedb were to be lost and then a client attempted
to modify a mutable file before the accounting crawler restored the
leasedb state for that file's shares, the write would fail with a
leasedb error.

Now the mutable file modification code re-initializes the leasedb state
relevant to the shares being modified, allowing the modification to
succeed whether or not there was leasedb state pertaining to those
shares at that point or not.

Fixes https://tahoe-lafs.org/trac/tahoe-lafs/ticket/2204